Create helper function to update JWT
